+++ src/cljstron/browser/plug.cljs
@@ -1,2 +1,0 @@
-(ns cljstron.browser.plug)
-