+++ src/cljstron/common/app.cljs
@@ -1,17 +1,17 @@
-(ns cljstron.browser.app
+(ns cljstron.common.app
   (:require
     [cljs.nodejs :as node :refer [enable-util-print!]]
     [clojure.tools.reader.edn :refer [read-string]]
     [cljs-node-io.core :as io :refer [slurp spit]]
     [cljs.pprint :refer [pprint]]
+    [clojure.string :as str]
     [cljstron.common.utils :refer [error]]
-    [cljstron.common.interop :refer [rjs->clj]]))
+    [cljstron.common.cljjs :refer [rjs->clj]]))
 
 (enable-util-print!)
 
 (defn- read-edn* [file]
   "Read `file` and tranform it from :edn to clojure structure."
-  (println "fichier : " file)
   (read-string (slurp file)))
 
 (def ^:export read-edn (memoize read-edn*))
@@ -28,11 +28,19 @@
   "Write`data` on file `file` in JSON format."
   (spit file (.stringify js/JSON (clj->js data) nil "  ")))
 
+(def key-functions #{:loop :symbol})
+
 ; forward
 (declare ^:private change-value)
 
 (defn- update-vector [conf value]
   (into [] (map (partial change-value conf) value)))
+
+; forward
+(declare ^:private function-change-value)
+
+(defn- function-update-vector [conf value]
+  (into [] (map (partial function-change-value conf) value)))
 
 ; forward
 (declare ^:private final-change-value)
@@ -41,36 +49,39 @@
   (into [] (map (partial final-change-value conf) value)))
 
 (defn- update-list [conf value]
-  (if (= (first value) :loop)
+  (if (get key-functions (first value))
     value
     (let [changed-value (list* (map (partial change-value conf) value))]
-      (if (and (= changed-value value)(not= (first value) :loop))
+      (if (= changed-value value)
         (error "(update-list) Non string value in list " changed-value)
         (if (every? string? changed-value)
           (apply str changed-value)
           changed-value)))))
 
 ; forward
-(declare read-app-edn)
+(declare read-app-edn2)
 
-(defn- loop-key [conf [key-list body]]
+(defn- symbol-key [conf symb & rest]
+  (let [symb (change-value conf symb)]
+    (apply symbol (str/split symb #"/"))))
+
+(defn- loop-key [conf [key-list body &rest]]
   (loop [ keys (first (seq (if (keyword? key-list)
                               (change-value conf key-list)
-                              (js/Error. ":loop keylist should be a key defined in :app"))))
+                              (error ":loop keylist should be a key defined in :app"))))
           res {}]
     (if (seq keys)
       (let  [ key (first (seq keys))
-              ; just to verify where we are... ;)
-              _ (println "key " key)
-              conf (read-app-edn key ".")
+              conf (read-app-edn2 key ".")
               val (change-value conf body)
-              val (final-change-value conf val)]
+              val (function-change-value conf val)]
         (recur (rest keys) (merge res {key val})))
       res)))
 
 (defn- final-update-list [conf value]
   (cond
     (= (first value) :loop) (loop-key conf (rest value))
+    (= (first value) :symbol) (symbol-key conf (first (rest value)))
     :else (error "(final-update-list) Non function in final list " value)))
 
 ; forward
@@ -80,6 +91,12 @@
   (into {} (map (partial update-kv conf) value)))
 
 ; forward
+(declare ^:private function-update-kv)
+
+(defn- function-update-map [conf value]
+  (into {} (map (partial function-update-kv conf) value)))
+
+; forward
 (declare ^:private final-update-kv)
 
 (defn- final-update-map [conf value]
@@ -87,14 +104,13 @@
 
 (defn- update-keyword [conf value]
   (cond
-    (= value :loop) value
+    (get key-functions value) value
     :else
       (if-let [new-value (get conf value)]
         new-value
         (error "The key '" value "' has no definition"))))
 
 (defn- change-value [conf value]
-  (println "change value : " value)
   (cond
     (vector? value)   (update-vector conf value)
     (list? value)     (update-list conf value)
@@ -102,12 +118,18 @@
     (keyword? value)  (update-keyword conf value)
     :else             value))
 
-(defn- final-change-value [conf value]
-  (println "final change value : " value)
+(defn- function-change-value [conf value]
   (cond
-    (vector? value)   (final-update-vector conf value)
+    (vector? value)   (function-update-vector conf value)
     (list? value)     (final-update-list conf value)
     ; (set? value)      (first value)
+    (map? value)      (function-update-map conf value)
+    :else             value))
+
+(defn- final-change-value [conf value]
+  (cond
+    (vector? value)   (final-update-vector conf value)
+    (set? value)      (first value)
     (map? value)      (final-update-map conf value)
     :else             value))
 
@@ -115,6 +137,11 @@
   (if (or (= key :parent) (= key :output))
     [key value]
     [key (change-value conf value)]))
+
+(defn- function-update-kv [conf [key value]]
+  (if (or (= key :parent) (= key :output))
+    [key value]
+    [key (function-change-value conf value)]))
 
 (defn- final-update-kv [conf [key value]]
   (if (or (= key :parent) (= key :output))
@@ -135,22 +162,20 @@
       (merge (merge-parents app parent) val)
       val)))
 
-(defn- read-app-edn* [key path]
-  (println "path = " path)
-  (let
-      ; read full config (TODO search cljstron directory if we are elsewhere)
-    [ app (load-def path)
-      ; merge value in parent data
-      config (merge-parents app key)]
-    (pprint config)
+(defn- read-app-edn2* [key path]
+  (let [config (merge-parents (load-def path) key)]
     (loop [conf config]
       (let [new-conf (into {} (map (partial update-kv conf) conf))]
         (if (= new-conf conf)
           ; finish : suppress sets, execute functions on data
-          (into {} (map (partial final-update-kv conf) conf))
+          (into {} (map (partial function-update-kv conf) conf))
           (recur new-conf))))))
 
-(def ^:export read-app-edn (memoize read-app-edn*))
+(defn ^:export read-app-edn [key path]
+  (let [conf (read-app-edn2 key path)]
+    (into {} (map (partial final-update-kv conf) conf))))
+
+(def ^:private read-app-edn2 (memoize read-app-edn2*))
 
 (defn ^:export write-gen-files [path]
   (loop [ ext (seq (:code (read-app-edn :-gen-files ".")))]
@@ -158,13 +183,13 @@
       (let [[ext-name ext-content] (first ext)
             output (:output ext-content)]
         (loop [ file (seq (dissoc ext-content :output))]
-          (println "\next-name " ext-name "\next-content " ext-content "\nfile " file "\noutput " output)
           (when file
-            (let [[filename content] (first file)]
-              (println "\nfilename " filename "\ncontent " content)
+            (let [[filename content] (first file)
+                  filename (str (name filename) "." (name ext-name))]
+              (println "generate " filename)
               (cond
-                (= output :edn)(write-edn (str (name filename) "." (name ext-name)) content)
-                (= output :json)(write-json (str (name filename) "." (name ext-name)) content)
+                (= output :edn)(write-edn filename content)
+                (= output :json)(write-json filename content)
                 :else (error "(write-gen-files) File type unknown" output))
               (recur (seq (rest file))))))
         (recur (seq (rest ext)))))))